added a bunch of user stuff

+++ backend/controllers/feed.controllers.js
@@ -1,9 +1,10 @@
 import { Post } from '../models/Post'
+import { User } from '../models/User'
 
 export const getAllPosts = (req, res, next) => {
     // Pagination
     const currentPage = req.query.page || 1
-    const perPage = 10
+    const perPage = 12
     let totalItems;
 
     Post.find()
@@ -32,18 +33,30 @@ export const savePost = (req, res, next) => {
     const title = req.body.title
     const content = req.body.content
 
+    let creator 
+
     const post = new Post({
         title: title,
         content: content,
-        creator: { name: 'Isaac' }
+        creator: req.userId
     })
 
     post
     .save()
     .then(result => {
+        return User.findById(req.userId)
+    }).then(user => {
+        creator = user
+        user.posts.push(post)
+        return user.save()
+    }).then(result =>{
         res.status(201).json({
             message: 'Post created successfully!',
-            post: result
+            post: post,
+            creator: { 
+                _id: creator._id, 
+                name: creator.name 
+            }
         });
     })
     .catch(err => {
@@ -79,6 +92,18 @@ export const updatePost = (req, res, next) => {
 
     Post.findById(postId)
     .then(post => {
+        if (!post) {
+            const error = new Error('Could not find post.')
+            error.statusCode = 404
+            throw error
+        }
+        if (post.creator.toString() !== req.userId) {
+            console.log('not authorized')
+            const error = new Error('Not Authorized')
+            error.statusCode = 403
+            throw error
+        }
+
         post.title = title
         post.content = content
         return post.save()
@@ -97,11 +122,33 @@ export const updatePost = (req, res, next) => {
 
 export const deleteOnePost = (req, res, next) => {
     const postId = req.params.postId
-
+    
     Post.findById(postId)
     .then(post =>{
+        //Check to make sure a post exists
+        if (!post ) {
+            const error = new Error('Could not find post')
+            error.statusCode = 404
+            throw error
+        }
+        //Check to see if the user trying to delete is the same user who created the post
+        if (post.creator.toString() !== req.userId) {
+            console.log('not authorized')
+            const error = new Error('Not Authorized')
+            error.statusCode = 403
+            throw error
+        }
         //check logged in user
         return Post.findByIdAndRemove(postId, { useFindAndModify: false })
+    })
+    .then(result => {
+        // Clear the relation between the post and the user
+        return User.findById(req.userId)
+        
+    })
+    .then(user => {
+        user.posts.pull(postId)
+        return user.save()        
     })
     .then(result => {
         res.status(200).json({
